buspasspro/buspasspro/backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dateutil==2.8.2
Pillow==11.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
razorpay==1.4.2